[dist-8x8] Do not use partitions or tx size smaller than 8x8
Because of the incompleteness, unreliability and difficult maintainability of the
code, CONFIG_DIST_8X8 experiment enforces RDO mode decision not to use sub 8x8 blocks
and tx smaller than 8x8 in RDO mode decision.

If --tune=[daala-dist|cdef-dist] is given, min partition size is enforced
to 8x8 pixels (4x16 and 16x4 are excluded as well) and tx_size < 8x8 is
excluded during RDO search.

To see the performance change of "--tune=[daala-dist|cdef-dist]",
compare with --enable-dist-8x8, which have the same mode decision options but
RDO uses PSNR as a distortion metric.

Change-Id: If052bd66f82a01e6f3675e9202e71f878b8baf5a